CHROMIUM: vulkan/runtime: Wait on Android acquire fence on CPU.
This works around stuck fences. Waiting on the CPU here can be a
"Quality of Implementation" issue and might result in lower perf.

However, this is overall very unlikely and this is the same behavior
as the Intel driver does so this is very low risk.

Ported for Mesa-22.0.3 form
https://chromium-review.googlesource.com/c/chromiumos/third_party/mesa/+/3158739/2

BUG=b:245687382,b:237232102,b:194858164,b:199406573,b:207433028
TEST=Pass CQ dry run

Change-Id: I6e7c3ac181d0fef524028c73c035484030fa9e8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/mesa/+/4082611
Commit-Queue: Ilja Friedel <ihf@chromium.org>
Reviewed-by: Ilja Friedel <ihf@chromium.org>
Tested-by: Ikshwaku Chauhan <ikshwaku.chauhan@amd.corp-partner.google.com>